购买商品从校园卡里扣费

+++ src/main/java/com/dj/ssm/pojo/Shop.java
@@ -7,6 +7,7 @@ import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 import lombok.experimental.Accessors;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -19,7 +20,7 @@ public class Shop {
 
     private String shopName;
 
-    private Double shopPrice;
+    private BigDecimal shopPrice;
 
     private Integer shopStatus;
